+++ demos/ros_examples/scripts/joint_space_velocity_control.cpp
@@ -4,7 +4,7 @@
 #include <sensor_msgs/JointState.h>
 #include <state_representation/space/cartesian/CartesianPose.hpp>
 #include <state_representation/space/cartesian/CartesianTwist.hpp>
-#include <state_representation/robot/JointVelocities.hpp>
+#include <state_representation/space/joint/JointVelocities.hpp>
 #include "dynamical_systems/DynamicalSystemFactory.hpp"
 #include <robot_model/Model.hpp>
