+++ protocol/clproto_cpp/include/clproto/encoders.h
@@ -6,7 +6,7 @@
 #include <state_representation/space/SpatialState.hpp>
 #include <state_representation/space/cartesian/CartesianState.hpp>
 #include <state_representation/space/Jacobian.hpp>
-#include <state_representation/robot/JointState.hpp>
+#include <state_representation/space/joint/JointState.hpp>
 #include <state_representation/parameters/Parameter.hpp>
 
 #include "clproto.h"
