+++ protocol/clproto_cpp/test/tests/test_json.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 
 #include <state_representation/space/cartesian/CartesianState.hpp>
-#include <state_representation/robot/JointState.hpp>
+#include <state_representation/space/joint/JointState.hpp>
 
 #include "clproto.h"
