+++ protocol/clproto_cpp/test/tests/test_packing.cpp
@@ -2,7 +2,7 @@
 
 #include <state_representation/space/cartesian/CartesianState.hpp>
 #include <state_representation/space/Jacobian.hpp>
-#include <state_representation/robot/JointState.hpp>
+#include <state_representation/space/joint/JointState.hpp>
 
 #include "clproto.h"
