+++ python/source/clproto/bind_clproto.cpp
@@ -10,11 +10,11 @@
 #include <state_representation/space/cartesian/CartesianTwist.hpp>
 #include <state_representation/space/cartesian/CartesianWrench.hpp>
 #include <state_representation/space/Jacobian.hpp>
-#include <state_representation/robot/JointState.hpp>
-#include <state_representation/robot/JointPositions.hpp>
-#include <state_representation/robot/JointVelocities.hpp>
-#include <state_representation/robot/JointAccelerations.hpp>
-#include <state_representation/robot/JointTorques.hpp>
+#include <state_representation/space/joint/JointState.hpp>
+#include <state_representation/space/joint/JointPositions.hpp>
+#include <state_representation/space/joint/JointVelocities.hpp>
+#include <state_representation/space/joint/JointAccelerations.hpp>
+#include <state_representation/space/joint/JointTorques.hpp>
 #include <state_representation/geometry/Shape.hpp>
 #include <state_representation/geometry/Ellipsoid.hpp>
