+++ source/controllers/src/impedance/VelocityImpedance.cpp
@@ -1,7 +1,7 @@
 #include "controllers/impedance/VelocityImpedance.hpp"
-#include "state_representation/robot/JointState.hpp"
-#include "state_representation/robot/JointPositions.hpp"
-#include "state_representation/robot/JointVelocities.hpp"
+#include "state_representation/space/joint/JointState.hpp"
+#include "state_representation/space/joint/JointPositions.hpp"
+#include "state_representation/space/joint/JointVelocities.hpp"
 #include "state_representation/space/cartesian/CartesianState.hpp"
 #include "state_representation/space/cartesian/CartesianPose.hpp"
 #include "state_representation/space/cartesian/CartesianTwist.hpp"
