+++ source/controllers/test/tests/test_cartesian_twist_controller.cpp
@@ -3,7 +3,7 @@
 #include "controllers/impedance/CartesianTwistController.hpp"
 #include "state_representation/space/cartesian/CartesianWrench.hpp"
 #include "state_representation/space/cartesian/CartesianTwist.hpp"
-#include "state_representation/robot/JointTorques.hpp"
+#include "state_representation/space/joint/JointTorques.hpp"
 
 using namespace controllers::impedance;
 using namespace state_representation;
