+++ source/state_representation/include/state_representation/space/Jacobian.hpp
@@ -2,8 +2,8 @@
 
 #include "state_representation/State.hpp"
 #include "state_representation/exceptions/IncompatibleSizeException.hpp"
-#include "state_representation/robot/JointTorques.hpp"
-#include "state_representation/robot/JointVelocities.hpp"
+#include "state_representation/space/joint/JointTorques.hpp"
+#include "state_representation/space/joint/JointVelocities.hpp"
 #include "state_representation/space/cartesian/CartesianTwist.hpp"
 #include "state_representation/space/cartesian/CartesianWrench.hpp"
