+++ source/state_representation/include/state_representation/space/joint/JointPositions.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "state_representation/robot/JointState.hpp"
-#include "state_representation/robot/JointVelocities.hpp"
+#include "state_representation/space/joint/JointState.hpp"
+#include "state_representation/space/joint/JointVelocities.hpp"
 
 namespace state_representation {
 class JointVelocities;
