+++ source/state_representation/src/space/joint/JointVelocities.cpp
@@ -1,4 +1,4 @@
-#include "state_representation/robot/JointVelocities.hpp"
+#include "state_representation/space/joint/JointVelocities.hpp"
 #include "state_representation/exceptions/EmptyStateException.hpp"
 
 using namespace state_representation::exceptions;
