+++ source/state_representation/test/tests/robot/test_joint_accelerations.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include "state_representation/robot/JointAccelerations.hpp"
+#include "state_representation/space/joint/JointAccelerations.hpp"
 #include "state_representation/exceptions/EmptyStateException.hpp"
 
 using namespace state_representation;
