+++ source/state_representation/test/tests/test_trajectory.cpp
@@ -3,7 +3,7 @@
 #include <unistd.h>
 #include "state_representation/trajectories/Trajectory.hpp"
 #include "state_representation/space/cartesian/CartesianState.hpp"
-#include "state_representation/robot/JointState.hpp"
+#include "state_representation/space/joint/JointState.hpp"
 
 TEST(TrajectoryTest, CreateTrajectory) {
   state_representation::Trajectory<state_representation::JointState> trajectory;
